wip on upgrade to schnorr
