+++ musig/Cargo.toml
@@ -11,3 +11,6 @@ rand = "0.6"
 subtle = "2"
 curve25519-dalek = { version = "1.0.1", features = ["serde"] }
 serde = { version = "1.0", features=["derive"] }
+
+[dependencies.schnorr]
+path = "../schnorr"
